refactor: addresed comments
