+++ cypress-tests/cypress/e2e/saved-card-flow-e2e-test.cy.ts
@@ -10,15 +10,12 @@ describe("Card payment flow test", () => {
   let getIframeBody: () => Cypress.Chainable<JQuery<HTMLBodyElement>>;
   let iframeSelector =
     "#orca-payment-element-iframeRef-orca-elements-payment-element-payment-element";
+     changeObjectKeyValue(createPaymentBody,"customer_id","hyperswitch_sdk_demo_id")
 
-  changeObjectKeyValue(createPaymentBody,"customer_id","hyperswitch_sdk_demo_id")
-
-  
   beforeEach(() => {
     getIframeBody = () => cy.iframe(iframeSelector);
     cy.createPaymentIntent(secretKey, createPaymentBody).then(() => {
       cy.getGlobalState("clientSecret").then((clientSecret) => {
-
         cy.visit(getClientURL(clientSecret, publishableKey));
       });
