+++ cypress-tests/cypress/e2e/stripe-manual-capture-card-flow-e2e-test.cy.ts
@@ -20,8 +20,7 @@ describe("Card payment flow test", () => {
     getIframeBody = () => cy.iframe(iframeSelector);
     cy.createPaymentIntent(secretKey, createPaymentBody).then(() => {
       cy.getGlobalState("clientSecret").then((clientSecret) => {
-
-        cy.visit(getClientURL(clientSecret, publishableKey));
+      cy.visit(getClientURL(clientSecret, publishableKey));
       });
 
     })
@@ -31,16 +30,13 @@ it("should complete the card payment successfully", () => {
   
     const { cardNo, card_exp_month, card_exp_year, cvc } = stripeCards.successCard;
 
-    // Wait for iframe to load and get its body
-    getIframeBody().find('[data-testid=cardNoInput]').type(cardNo); // Example card number
-    getIframeBody().find('[data-testid=expiryInput]').type(card_exp_month); // Expiration month
-    getIframeBody().find('[data-testid=expiryInput]').type(card_exp_year); // Expiration year
-    getIframeBody().find('[data-testid=cvvInput]').type(cvc); // CVV
+    getIframeBody().find('[data-testid=cardNoInput]').type(cardNo); 
+    getIframeBody().find('[data-testid=expiryInput]').type(card_exp_month);
+    getIframeBody().find('[data-testid=expiryInput]').type(card_exp_year); 
+    getIframeBody().find('[data-testid=cvvInput]').type(cvc);
 
-    // Click on the submit button
     getIframeBody().get("#submit").click();
-
-     // Wait for the response and assert payment processing message
+    
   cy.wait(3000);
   cy.contains('Payment processing! Requires manual capture')
   .should('be.visible');
